Made validations on requests

+++ app/Http/Controllers/ProductController.php
@@ -4,9 +4,22 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Product;
+use Illuminate\Support\Facades\Validator;
 
 class ProductController extends Controller
 {
+    protected function validationTreatment(Request $request){
+        return $validator = Validator::make($request->all(),[
+        'name' => 'required',
+        'value' => 'required|numeric|min:0|regex:/^\d+(\.\d{1,2})?$/',
+        'description' =>''
+    ],[
+        'name.required' => 'É necessário um nome.',
+        'value.required' => 'É necessário um valor.',
+        'value.numeric' => 'Valor deve ser numérico.',
+        'value.min' => 'Valor deve não deve ser negativo.',
+        'value.regex' => 'Valor deve conter apenas duas casas decimais']);
+    }
     /**
      * Display a listing of the resource.
      */
@@ -20,6 +33,11 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
+        
+        $validator = $this->validationTreatment($request);
+        if($validator->fails()){
+            return response()->json(['message' => $validator->errors()->first()], 400);
+        }
         return Product::create($request->all());
     }
 
@@ -28,7 +46,11 @@ class ProductController extends Controller
      */
     public function show(string $id)
     {
-        return Product::find($id);
+        $product = Product::find($id);
+        if($product){
+            return $product;
+        }
+        return response()->json(['message'=> 'Produto não encontrado'], 404);
     }
 
     /**
@@ -36,9 +58,16 @@ class ProductController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $product = Product::findOrfail($id);
-        $product->update($request->all());
-        return $product;
+        $product = Product::find($id);
+        if($product){
+            $validator = $this->validationTreatment($request);
+            if($validator->fails()){
+                return response()->json(['message' => $validator->errors()->first()], 400);
+            }
+            $product->update($request->all());
+            return $product;
+        } 
+        return response()->json(['message'=> 'Produto não encontrado'], 404);
     }
 
     /**
@@ -46,7 +75,13 @@ class ProductController extends Controller
      */
     public function destroy(string $id)
     {
-        Product::destroy($id);
-        return "produto excluído com sucesso";
+        $product = Product::find($id);
+        if($product){
+            Product::destroy($id);
+            return response()->json(['message'=> 'produto excluído com sucesso'], 200);
+        }
+        
+        return response()->json(['message'=> 'Produto não encontrado'], 404);
+
     }
 }
